LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focusgroup/tentative/grid-navigation/row-wrap-and-col-flow-expected.txt

r1c1	r1c2	r1c3
r2c1	r2c2	r2c3

FAIL When on the last column, a right arrow press should move the focus to first column if the focusgroup wraps horizontally. assert_equals: expected Element node <td id="r1c1" tabindex="0">r1c1</td> but got Element node <td id="r1c3" tabindex="-1">r1c3</td>
FAIL When on the last row, a down arrow press should move the focus to first row and next column if the focusgroup flows vertically. assert_equals: expected Element node <td id="r1c2" tabindex="-1">r1c2</td> but got Element node <td id="r2c1" tabindex="-1">r2c1</td>
FAIL When on the first column, a left arrow press should move the focus to last column if the focusgroup wraps horizontally. assert_equals: expected Element node <td id="r1c3" tabindex="-1">r1c3</td> but got Element node <td id="r1c1" tabindex="0">r1c1</td>
FAIL When on the first row, an up arrow press should move the focus to last row and previous column if the focusgroup flows horizontally. assert_equals: expected Element node <td id="r2c3" tabindex="-1">r2c3</td> but got Element node <td id="r1c1" tabindex="0">r1c1</td>
